Move vars and nodeset variables to project definition
The nodeset definition and variables that were set in the
edpm-ansible-github-rdo-integration-centos-8-crc-singlenode-centos-9-external-compute
job definition might be problematic for other Zuul that is taking
config repository and the nodeset is not created, so tt requires to
create fake nodeset.
Moving nodeset and variables to the project definition remove that
problem.

Change-Id: I1e4cdf082803b302a304be5dc8be457dc0bc5c4e