Add job definition for data_plane_adoption job and nodeset
This adds data-plane-adoption-github-rdo-centos-8-crc-centos-9-standalone
The parent is added in [1] and it inherits from the upstream
standalone-standard. The pre plays in the parent job
install and setup crc and the pre in this job will run the
use_install_yamls role to set up the podified ctlplane. Note we do not
override the run: play since we also need a standalone deployment, ie
the install_yamls setup is passed via pre-run:

Also adds the new centos-8-crc-single-node-centos-9-standalone
nodeset

Testing at [2]

[1] https://review.rdoproject.org/r/c/config/+/46876
[2] https://review.rdoproject.org/r/c/testproject/+/46923/3#message-2b1cb0b24143ec3c3eaeb3c42232822211730029

Change-Id: If218df886b5a189f9d59ee987e3a812d50e282cf