WIP adding job definition for data_plane_adoption job and nodeset
This adds data-plane-adoption-github-rdo-centos-8-crc-centos-9-standalone
The parent is added in the depends-on and it inherits from the
upstream standalone-standard. The pre plays in the parent job
install and setup crc and the pre in this job will run the
use_install_yamls role to set up the podified ctlplane. Note we do not
override the run: play since we also need a standalone deployment, ie
the install_yamls setup is passed via pre-run:

Also adds the new centos-8-crc-single-node-centos-9-standalone
nodeset

Depends-On: https://review.rdoproject.org/r/c/config/+/46876

Change-Id: If218df886b5a189f9d59ee987e3a812d50e282cf