Use edpm-ansible last commit as a tag for openstack-runner
Currently we were setting the tag based on openstack-ansibleee-operator
repo tag.
The zuul job builds the openstack-runner image but does not pushes
the image as the image having ansibleee-operator last commit as a
tag exists in the registry. It fails to test the correct content.

This patch fixes the same to use the edpm-ansible last commit
as a tag.

Signed-off-by: Chandan Kumar <chkumar@redhat.com>
Change-Id: I4c2262b0fdc9bdf44318d538563e35e9d62324cb