Disable execution of validations catalog tests on master and enable them only in wallaby
Issue with keystone token retrieval has caused several validations to fail
in master branch promotion pipeline. In absence of keystone fix,
or alternative method of token retrieval, disabling tests
of the validations affected is the only way to unblock promotion pipeline.

This change should be reverted if the issue is resolved.

Affected validations:
- overcloud-service-status
- openstack-endpoints
- openshift-nw-requirements

Closes-Bug: #2006125

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I3812f5e12c082cc132df759a3b82e3a5f66120a4