Clean some of the playbook
With If96beb8bd2285d4de7c96c2baa5f3636512a8827 we should be able to not
build the ansibleee related containers in this playbook as they should
be built during the prep steps.

We also need to ensure openstack-k8s-operators/openstack-operator is
present soon enough, so we add it to the direct dependencies that must
be deployed during the workspace preparation.

While the "config" repository is trusted and, therefore, can't get any
depends-on on on patches therein, the following reference will ensure we
can follow the full process:

Depends-On: https://review.rdoproject.org/r/c/config/+/47002
Change-Id: I2d6cccf61ce5606862eb09c5814f223cde357d27