Consolidate weirdo playbooks
In order to avoid having to apply the same settings in
all weirdo playbooks, we put the common tasks in the same
file.
We don't put them in a "pre-run" playbook as there are "weirdo"
tasks, so they need to be run within the "run" playbook if we
follow the Ansible philosophy.

Change-Id: I5b28722d4fd0f34a62fdc721edee8e712424e3ea