Add Manila in openstack CR
Manila was recently integrated in Meta operator with [1].
We use a custom CR is our jobs, we need to update our CR
to add Manila. This will fix the below issue which we are
seeing in our jobs. More details in [2].

~~~
The OpenStackControlPlane "openstack" is invalid:
spec.manila.template.manilaShares: Invalid value: "null":
spec.manila.template.manilaShares in body must be of type object: "
~~~

[1] https://github.com/fao89/openstack-operator/commit/a822b6fbb8e05f739f0606b0ddef5f2acf6ad708
[2] https://logserver.rdoproject.org/23/46923/16/check/data-plane-adoption-github-rdo-centos-8-crc-centos-9-standalone/1ee17f9/job-output.txt

Change-Id: Ifbe153b5d6dd69835c43a8b367a346f9e17f270f