Add standalone IP, mariadb pass, ssh key for mariadb dpa test
This slurps the ip address of the standalone node and the password
for mariadb and sets the values into the vars and secrets.yaml
files used by the data-plane-adoption mariadb_copy role.

Also setup iptables forwarding to allow crc node access to mariadb and
adds the crc zuul-worker public key to the standalone authorized_keys
to allow ssh access for stopping openstack services.

Change-Id: I73ae112e43a613c57c720bfda2e25335a992d71c