Create two non-voting jobs to compare requires and BRs from latest trunk
RDO is moving specs to the new pyproject-rpms-macros. Part of this
change is automating the runtime and buldtime requirements. In order to
observe the changes en dependencies produced while automating
dependencies, this patch is creating two new jobs:

- DLRN-rpmbuild-centos9-compare-deps: compare runtime dependencies from
  the resulting rpms with the same packages from latest RDO Trunk.
- DLRN-rpmbuild-centos9-compare-builddeps: compare buildtime
  dependencies from the srpm created by dlrn with the last srpm from RDO
  Trunk.

This patch is assigning these jobs to rpm-master branch of all the
openstack packages as non-voting. I'm assuming some diferences may be
acceptable or even required but we want to monitor those changes before
merging the patches.

Change-Id: I1a73dad06c059a104ceefdb3d8be61406af9a0e6