Install osclient, set etc.hosts and passwords for keystone glance placement dpa
This installs openstackclient and adds an entry for podified keystone
glance and placement into the etc hosts pointing to the crc VM.
Also tweaks the admin password used in the tests to correct values.

This also grabs the glance and placement passwords from the tripleo
passwords file and sets it in secrets.yaml used by the dpa tests. It
removes the unneeded set_fact for the tripleo passwords.

Change-Id: I3eec98d42b427b459199ea50895537125cd312d9