Tweak the failed_when for the dataplane adoption tests and add logs
This tweaks the failed_when condition that detects when the dpa
test play fails. It also adds some more files in the collect logs
tasks for the job vars/secrets etc. It also sets the use_no_log
parameter to false so we get the full logs (there are no secrets
to be leaked in this job).

Change-Id: I82d8916c5057a531fb179f7d87c5d6ac72921593