Speed up Adoption job by not deploying control plane before Adoption
Previously the job first deployed podified control plane, then tore it
down, and then ran the adoption procedure. It should be fine to just
run the adoption procedure straight away. That means we won't need to
reset the CRC storage in the test suite. Both of these changes will
result in speeding up the job.

Change-Id: I628e2000147d1e09bdffd9336c22abcfcb5c00ec