Stop deploying the podified controlplane before the dpa tests
The test suite starts by removing the existing deployment before
deploying the podified ctlplane for the test suite. This removes
the deployment before the dpa tests are executed. We also tweak
the crc params to pass 8 cpus for the crc vm.

Also adds rabbit and ocp to pre_pull_images to avoid errors
pulling during the deployment

Change-Id: Iac2ed11cae5d4631f4e307c66a616208175e18ff