Create new adoption job that runs in a single host
Add a new adoption job that runs in a single large virthost. Inside this
host we create two vms, one for crc and one to deploy standalone, with
network isolation in both environments.

Add a role to run preparation commands in the vm before the tripleo
deployment, and modify the data-plane adoption job to configure the
standstandalone deployment for network isolation.

Change-Id: Iaa425e5cda744908be7dd46c444772ef3a046a4b