Use dnf repoquery for compare-deps job
We are hitting similar issue as we had it compare-files
job [1]

Also, we install the packages from Trunk repos one by one
in order to not make the whole process fail if one of the
packages is not available (e.g extra subpkgs added but not
in Trunk repo). As we don't mock --init after each installation
we get the list of deps installed (rpm -qa) of all successful
packages installed at the end.

[1] https://review.rdoproject.org/r/c/rdo-jobs/+/48993

Change-Id: Ia95e00f9f7b4cdb2431a52f0681f03f94173a1a9