Handle extras subpkg in compare jobs
In compare-deps jobs, we install the packages from Trunk repos one
by one in order to not make the whole process fail if one of the
packages is not available (e.g extra subpkgs added but not
in Trunk repo). As we don't mock --init after each installation from
Trunk, we get the list of deps installed (rpm -qa) of all successful
packages installed at the end.

In compare-files job, we compare the packages only when they both
exist in Trunk and patched repos.

Also, we are hitting similar issue as we had in compare-files
job [1], so we switch to 'dnf repoquery'.

[1] https://review.rdoproject.org/r/c/rdo-jobs/+/48993

Change-Id: Ia95e00f9f7b4cdb2431a52f0681f03f94173a1a9