[WIP] Install CA for openstack TLS in dpa single-node
Recently, openstack has enabled TLS support, this change installs the CA
into the single-node data-plane adoption job. This is copied from the
ci-framework edpm_prepare role, but since the single-node job is going
away soon, it's faster to just duplicate the few tasks than moving to
job to consuming further ci-framework.

To get the CA data, we need to create a dummy openstack deployment that
we then delete after gathering the CA.

Change-Id: Ifb836c164d7c8cce6ed607bcc0c9113efd424c0e