Deploy dummy controlplane in openstack namespace
In the single node adoption job we need to deploy a dummy controle plane
to extract the CA from the secret that contains it. Until now it did not
specify any namespace when deploy the control plane, but lately it has
begun failing because it's deploying in a namespace different than
openstack.

Change-Id: I3668e683b804112af07c0d53db0a3b139ca72123