Adding new data-plane-adoption-osp17-multinode-to-extracted-crc
As part of [1] adds the job definition and plays. For now we are adding
undercloud installation with the Depends-On. Followup patches will add
the overcloud deployment but we can merge this to allow us to proceed
in step-wise manner. The job is not wired up to run anywhere yet.
Testing at [2].

https://issues.redhat.com/browse/OSPRH-4411

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/1028
Depends-On: https://github.com/openstack-k8s-operators/install_yamls/pull/693/

[1] https://issues.redhat.com/browse/OSPRH-3038
[2] https://review.rdoproject.org/r/c/testproject/+/51274/6#message-441fadff3c14cef8b21a7560afe6158d150fced1

Change-Id: I03d887d8378ed9adb6f9519609c61bc87c879d28