Adds overcloud os-net-config and ssh setup for new data plane adoption job
As part of [1] this adds the os-net-config run for the overcloud nodes
and setup ssh from the undercloud/director node.

Finally, this moves the standalone os-net-config setup back into its own
play (i.e. partial revert of [2]). We can try and recombine the
standalone and over/undercloud os-net-config setup later as an
improvement once this work is settled.

Depends-On: https://github.com/openstack-k8s-operators/install_yamls/pull/745
Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/1201

[1] https://issues.redhat.com/browse/OSPRH-3038
[2] https://review.rdoproject.org/r/c/rdo-jobs/+/51227/20/playbooks/data_plane_adoption/setup_tripleo_os_net_config.yaml

Change-Id: I403483acf14f2ff2505c7e598608be292cfbaf5f