Manage branch name change to unmaintained/yoga in weirdo jobs
Upstream puppet projects have moved to unmaintained/yoga branches as per
current unmaintained policy.

In weirdo jobs we are using rdopkg to find out the upstream branch names for
releases and rdopkg assumes stable/ prefix always [1].

Until we have a proper fix via change in rdoinfo/rdopkg, I am proposing
an as-hoc hack to workaround the issue for yoga.

[1] https://github.com/softwarefactory-project/rdopkg/blob/e5721d98c878af38179c0c265a0706e3872dc0b9/rdopkg/actionmods/rdoinfo.py#L54

so weirdo is falling back to master and failing.

Change-Id: Ic13d309f7fa93f4f23ad25c52951dca5117afd73