Wire up test-minimal data plane adoption test target multinode tripleo
Wires up the test-minimal target from the data plane adoption tests repo
for the new tripleo multinode job. The overcloud/tripleo deployment was
added in the parent patch [1] and its depenencies. This is part of [2].

The periodic-data-plane-adoption-multinode-osp-17-to-extracted-crc-no-ceph is
wired up to run next to the existing job, but not yet in promotion criteria.

https://issues.redhat.com/browse/OSPRH-5754

Depends-On: https://github.com/openstack-k8s-operators/data-plane-adoption/pull/397

[1] https://review.rdoproject.org/r/c/rdo-jobs/+/51916
[2] https://issues.redhat.com/browse/OSPRH-3038

Change-Id: I104399d783a46cd58b245a0aab0aa2e9ab6ffa0e