Move tox jobs to CentOS Stream 9 in rdo-jobs repo
As part of the migration of automation jobs out of cs8 anc c7 because of
EOL we need to move jobs running in CS8 to CS9 nodes.

This patch is moving tox-linters and tox-docs to CS9 nodes.

pbr is not longer supporting build_sphinx so I am also moving out of pbr
to build docs in tox.ini.

Change-Id: I0dfedfb7d6c19c7f00f8c17af460f0c2c4b39d22