Allow choosing test suite to run for multinode
Allow choosing which test suite to run for adoption multinode node jobs.
Defaults to test-minimal. Set the multinode with ceph job to run the
'test-with-ceph' test suite. Additionally, do not write success file if
'force_job_failure' is set to true, so that we can hold an env with
succesuccessful tests.

Change-Id: Ifd3d75b0c5a467bc844e8945155e98361e27d0ab