Add final two compute nodes for data plane adoption multinode job
This introduces two more compute nodes for the multinode tripleo
source adoption job. The edpm_computes and edpm_nodes are crafted to
include all compute before being passed in vars.yaml.

This also passes the source_galera_members var as it was previously only
taking into account the first controller (and with a wrong hostname for
the multinode case) to be in depends-on data-plane-adoption/pull/482.

https://issues.redhat.com/browse/OSPRH-5756

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/1802
Depends-On: https://github.com/openstack-k8s-operators/install_yamls/pull/840
Depends-On: https://github.com/openstack-k8s-operators/data-plane-adoption/pull/482

Change-Id: Ibeec3fa706900f5cced4cb3abfc89ca77e69c46a