Add validate-buildsys-tags-dalmatian-testing-cifmw-crc-podified-edpm-baremetal job
We set the job as non-voting for now, until we are sure it's
stable enough.

We are using the nodeset 'cloud-centos-9-stream-vexxhost'
which have specific firewall rule to accept connection to
the registry (i.e TCP/5001).
For now, we are using 4.14.0-0.okd-2023-12-01-225814 which is running on
Fedora CoreOS (FCOS) but in the future we'll test on newer OpenShift
release hosted on CentOS Stream CoreOS (SCOS).

Change-Id: I2c7ffc2a630a4b82fbdaf324f27e37758c3dd8b6