Add subscription manager repo disable before adoption tests execution
Lets disable the 17.1 repos before we run the adoption tests.
We need to ensure that after the tripleo installation and before
we setup the osp18 data plane we are not consuming any more 17.1
packages. This also fixes the unregister task in the multinode case
as it was not working as expected (inventory hostname not found).

Related: https://issues.redhat.com/browse/OSPRH-8437

Change-Id: I0c629bc00c20b10fcd466c074182d10d4076bd91