Update data plane adoption job names for consistency
So data-plane-adoption-osp-17-to-extracted-crc is now
adoption-standalone-to-crc-ceph and we have a consistent naming scheme.
This keeps the existing jobs around (and they parent on the 'new' job)
so we can merge this without breaking github-check jobs across the various
repos that are running adoption. There are related patches in those repos
to re-parent onto the new job that depends-on this patch [1][2][3][4].

Finally we will followup to remove the existing jobs altogether.

Related: https://issues.redhat.com/browse/OSPRH-8452

[1] https://github.com/openstack-k8s-operators/ci-framework/pull/2091
[2] https://github.com/openstack-k8s-operators/install_yamls/pull/872
[3] https://github.com/openstack-k8s-operators/edpm-ansible/pull/698
[4] https://github.com/openstack-k8s-operators/data-plane-adoption/pull/530

Change-Id: Ia0ac8c34728c4f6dd3f255095d6661c124b7688d