Adds task to merge main skiplist and exclude list
The cifmw_test_operator_tempest_exclude_list and the master skiplist do not work together. This patch adds tasks that will merge them together and send to the test-operator. As exclude list was not working last time, we could not add nova-operator tests as it has a bunch of known CIXs.
I am adding nova-operator to the test list to show that it indeeds considers the master exclude list and the job passes.


Change-Id: I1f8cded8c1782f5b80e79c6392ebc2a1402ce257