Rework adoption set facts and edpm_nodes for multinode networker job
This makes the block setting the edpm_nodes and related vars generic
so we can use it in the different multinode scenarios. It also fixes the
hostnamemap creation so we can control the hostnamemap/placement in a more
predictable way using the zuul groups we have from the zuul nodeset.

https://issues.redhat.com/browse/OSPRH-8810

Change-Id: I6673f60bc2d8b4073589ac3391981952a40f2a93