Move edpm_node_ips to zuul job vars from adoption playbook for multinode
As part of OSPRH-8810 this moves the edpm_node_ips from the adoption
deploy_tripleo_run_repo_tests.yaml vars: to the zuul job vars so it.
This is needed by [1] where we can rely on the edpm_node_ips to be
set correctly depending on the job type. Generic multinode has 3
compute but networker multinode has 2.

https://issues.redhat.com/browse/OSPRH-8810

[1] https://review.rdoproject.org/r/c/rdo-jobs/+/54031/4/playbooks/data_plane_adoption/deploy_tripleo_run_repo_tests.yaml#b183

Change-Id: I5a81561ac4500fe2ead1e2c8d70e6d2507ebb579