Improve the enablement and disablement of repos
Where recently hitting rate limits when enabling repos for tripleo
deployment. This change moves the enabling to do it all in one command.

Additionally, in adoption multinode jobs, we were using the use_ceph var,
but not casting to a bool, which would always evaluate to true as long as
the string is not empty. Add casting to bool so it does what was intended.

Change-Id: I8a7891a1fa084decbe0ac2e356e21df66c337815