Remove redundant openstack-deployment-tools for multinode adoption ci
This removes one of the repos enabled for the OSP17.1 deployment
in multinode adoption jobs. The repo must have switched names
or is otherwise now no longer available. Current jobs are passing
even though there is an error enabling this see [1][2] for recent
examples up/dstream respectively. There is a new deployment tools we
could add instead but the passing jobs mean it is redundant so lets
remove instead.

[1] https://logserver.rdoproject.org/96/53696/2/check/periodic-adoption-multinode-to-crc-ceph/47b5789/job-output.txt
[2] https://sf.hosted.upshift.rdu2.redhat.com/logs/openstack-periodic-integration-rhos-18-rhel9/gitlab.cee.redhat.com/ci-framework/ci-framework-jobs/main/periodic-internal-adoption-multinode-to-crc-ceph/eb27e1c/job-output.txt

Change-Id: I0ef69c1826036884ae110d48e59d7c4c5c0e95cf