Disable octavia for periodic adoption jobs
Since octavia adoption is not fully implemented yet, if we enable it in
the source environment when we run tempest it detects the load-balancer
service, but the actual service is stopped during adoption,so it
crahses. This patch disables octavia for the source env in the periodic
jobs, but keeps it in the check jobs to allow for development to
proceed.

Change-Id: I6d86c8193c93b7829fd826bff2b1783b6252bc23