Add hugepages coverage for noTLS ceph adoption job
Provide vars to deploy standalone and multinode adoption jobs
with hugepages.

When tripleo is deployed with hugepages support for
rhoso adoption testing, add tasks to reboot the standalone,
or a 1st compute host in a multinode setup, and to wait for
the nova compute service started. For that to work in multinode
setup, run test preparation playbook twice: in undercloud vm,
and overcloud by provided inventories.

Note: cannot use TLSe with hugepages together, because freeipa
container does not survive the host reboot.

Add hugepages coverage to
periodic-adoption-multinode-to-crc-no-ceph.

Jira: #OSPRH-10611

Change-Id: I4eb0e57b1b7b2a33afa5ef12a220aa9ce5852c78
Required-by: https://github.com/openstack-k8s-operators/install_yamls/pull/938