Increase timeout for multinode ceph adoption job
The multinode ceph job runs "test-with-ceph" test suite which takes around 5 hours to complete with the current list of steps we have. This patch increases the timeout to 5 hrs. 
We realized this when this job started failing with CIX: https://issues.redhat.com/browse/OSPCIX-557
The multinode no ceph job does not need this as it runs test-minimal test suite which takes less time. 
This is not needed in the standalone jobs too as they take less time (less nodes)

Change-Id: I2a22a2721c28528867ad7870e1bf8ddbf1857b4f