Add fake ubuntu-noble nodeset
As per [1], we need to define fake nodeset in order
to satisfy Zuul config including some devstack
jobs from upstream.

Also, we define a fake job in order to be able to run
devstack job as 3rd party CI job and report back to
upstream patches.

As a followup of [2][3]

[1] https://review.rdoproject.org/r/c/rdo-jobs/+/33577
[2] https://review.rdoproject.org/r/c/config/+/55831
[3] https://review.rdoproject.org/r/c/config/+/55832

Change-Id: I14aa6e12ad626332589cd418d85664012a5ad5f1