Run swift migration tests post-adoption
Use the existing set of tests from install_yamls.

Always enable swift_replicated in the base multinode non-ceph adoption
job, as the original environment has multiple swift services.

Finally, enable the swift migration tests in the periodic multinode
non-ceph job, as they adds just a few minutes to the runtime
and also they can be seen as the equivalent of the ceph-specific
update tests executed in the ceph job.

Change-Id: I75fca5211c164b06b23d16ee3e5a3f90a5d32153