Disable swift migration in no-ceph-rollback temporary
Swift migration is not currnetly possible to be tested in
periodic-adoption-multinode-to-crc-no-ceph-rollback as job
is testing adoption to 18 and later rollback backs to 17.1.

This is done in single stage and should be seperated into multiple
stages as improvment.[1]

Once seperation of adoption and rollback is done we can re-enable
swift migration in no-ceph-rollback job.Swift migration should happen
between adoption and rollback.

Resolves: OSPCIX-675

[1]https://issues.redhat.com/browse/OSPRH-12711

Change-Id: Idd3e1ceae069bd4c22c18d5384595729622169f9