Do not enable subcription and install packages via shell; move to modules
In the CI jobs, sometimes we see errors that it is just side effect
of early failed task, which has been ignored.
In that case, in the logs:

    TASK [Setup repos and install deps for running os-net-config]
    standalone | This system has no repositories available through subscriptions.
    standalone | This system has no repositories available through subscriptions.
    standalone | Updating Subscription Management repositories.
    standalone | Last metadata expiration check: 0:32:20 ago on Fri May 16 06:56:11 2025.
    standalone | No match for argument: openvswitch
    standalone | No match for argument: os-net-config
    standalone | Error: Unable to find a match: openvswitch os-net-config

The step that openvswitch can not be found might be ignored.
Let's move the commands executed in shell and use proper Ansible
modules.

Change-Id: I320f4a9f273976bbe99204061e4555d155931b63
Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>