Push cs10 master containers with current tag
Since there is no CS10 DLRN endpoint and cs10 EDPM nodeset.
In order to pull containers in EDPM job, we need to use proper
tag.

The current tcib job tags the images based on hashes.
https://github.com/openstack-k8s-operators/ci-framework/pull/2991
provides a way to retag the image with different named tag.

Let's use this one and re-tag the same image with current
tag.

This cr also drops exclude list as
https://github.com/openstack-k8s-operators/tcib/pull/274 fixes
the horizontests container.

Change-Id: I49bbd856f40dd86746f8d80528de318ff3978d4f
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>