Include CI Framework all.yml when nested Ansible; change basedir path
The nested Ansible execution did not read the group_vars/all.yml
variable file, so in the tasks some variable would be missing.

Also changed value for cifmw_repo_setup_basedir variable when
repo_setup role is called, due the path is unproper and it raises
complications in later stage.

Change-Id: I3428b1126f6b25d600d098e32680d1f9316fc56b
Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>